Fixes bug in previous commit.
