+++ config/initializers/dul_hydra.rb
@@ -10,12 +10,12 @@ DulHydra.configure do |config|
   config.remote_groups_name_filter = "duke:library:repository:ddr:"
   config.component_download_group = "duke:library:repository:ddr:component_download"
   config.extra_ability_logic = [:discover_permissions, 
-                              :export_sets_permissions, 
-                              :preservation_events_permissions,
-                              :batches_permissions,
-                              :ingest_folders_permissions,
-                              :download_permissions
-                             ]
+                                :export_sets_permissions, 
+                                :preservation_events_permissions,
+                                :batches_permissions,
+                                :ingest_folders_permissions,
+                                :download_permissions
+                               ]
   if File.exists? "#{Rails.root}/config/groups.yml"
     config.groups = YAML.load_file("#{Rails.root}/config/groups.yml")
   end
